Alternative solution using inner classes
