--- part10-Part10_17.SortThemCards/src/main/java/SortBySuit.java
@@ -1,19 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/**
- *
- * @author crist
- */
-import java.util.Comparator;
-
-public class SortBySuit implements Comparator<Card> {
-    
-    public int compare(Card c1, Card c2) {
-        return c1.getSuit().ordinal() - c2.getSuit().ordinal();
-    }
-    
-}
